AAI: Affordability changes for power plants
This changes the way power plants are selected, based on the observation
that AAI would start building advanced solars much too early, before they
are affordable.
The core of the change is to introduce a new method UnitAffordable that
can prevent units from being selected which cannot be reasonably built
given the current level of economy.  The selection algorithm is modified
a bit further; AAIExecute no longer gets to select a desired level of
efficiency.  The new algorithm is to select the most efficient affordable
power plant; however, there's still the old urgency bonus that could lead
to a smaller plant being selected.

This probably needs further tuning, but at least it now keeps building
regular solars at the start.