AAI: Reduce code duplication in GetAssault functions
There are five near-identical functions, one for each assault category.
Consolidate them into a single one that takes the category as argument.

A side effect is that we can now request ground/air/hover units that
are effective against submarines, and vice versa.  I can't see a reason
why there couldn't be a mod that allows this.